moved tests and updated openapi.json
