+++ services/dynamic-sidecar/tests/unit/test_api_containers.py
@@ -8,11 +8,14 @@ from typing import Any, Dict, Generator, List
 
 import pytest
 from async_asgi_testclient import TestClient
+from faker import Faker
 from fastapi import status
 from simcore_service_dynamic_sidecar._meta import api_vtag
 from simcore_service_dynamic_sidecar.settings import DynamicSidecarSettings
 from simcore_service_dynamic_sidecar.shared_handlers import write_file_and_run_command
 from simcore_service_dynamic_sidecar.shared_store import SharedStore
+
+DEFAULT_COMMAND_TIMEOUT = 10.0
 
 pytestmark = pytest.mark.asyncio
 
@@ -73,6 +76,74 @@ async def started_containers(test_client: TestClient, compose_spec: str) -> List
 @pytest.fixture
 def not_started_containers() -> List[str]:
     return [f"missing-container-{i}" for i in range(5)]
+
+
+async def test_compose_up(
+    test_client: TestClient, compose_spec: Dict[str, Any]
+) -> None:
+
+    response = await test_client.post(
+        f"/{api_vtag}/containers",
+        query_string=dict(command_timeout=DEFAULT_COMMAND_TIMEOUT),
+        data=compose_spec,
+    )
+    assert response.status_code == status.HTTP_201_CREATED, response.text
+    assert json.loads(response.text) is None
+
+
+async def test_compose_up_spec_not_provided(test_client: TestClient) -> None:
+    response = await test_client.post(
+        f"/{api_vtag}/containers",
+        query_string=dict(command_timeout=DEFAULT_COMMAND_TIMEOUT),
+    )
+    assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR, response.text
+    assert json.loads(response.text) == {"detail": "\nProvided yaml is not valid!"}
+
+
+async def test_compose_up_spec_invalid(test_client: TestClient) -> None:
+    invalid_compose_spec = Faker().text()
+    response = await test_client.post(
+        f"/{api_vtag}/containers",
+        query_string=dict(command_timeout=DEFAULT_COMMAND_TIMEOUT),
+        data=invalid_compose_spec,
+    )
+    assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR, response.text
+    assert "Provided yaml is not valid!" in response.text
+    # 28+ characters means the compos spec is also present in the error message
+    assert len(response.text) > 28
+
+
+async def test_containers_down_after_starting(
+    test_client: TestClient, compose_spec: Dict[str, Any]
+) -> None:
+    # store spec first
+    response = await test_client.post(
+        f"/{api_vtag}/containers",
+        query_string=dict(command_timeout=DEFAULT_COMMAND_TIMEOUT),
+        data=compose_spec,
+    )
+    assert response.status_code == status.HTTP_201_CREATED, response.text
+    assert json.loads(response.text) is None
+
+    response = await test_client.post(
+        f"/{api_vtag}/containers:down",
+        query_string=dict(command_timeout=DEFAULT_COMMAND_TIMEOUT),
+    )
+    assert response.status_code == status.HTTP_200_OK, response.text
+    assert response.text != ""
+
+
+async def test_containers_down_missing_spec(
+    test_client: TestClient, compose_spec: Dict[str, Any]
+) -> None:
+    response = await test_client.post(
+        f"/{api_vtag}/containers:down",
+        query_string=dict(command_timeout=DEFAULT_COMMAND_TIMEOUT),
+    )
+    assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR, response.text
+    assert json.loads(response.text) == {
+        "detail": "No spec for docker-compose down was found"
+    }
 
 
 async def test_containers_get(
